Fix wrong setting for previous node in dlist.rs
When inserting new node at head, current code sets the prev
pointer of the second to itself. So it should generate infinite-loop
if somebody reads the list reversely.

This patch changes unshift() function to set the prev pointer
of the second node to the first node.

I'm attaching a test code below. Before fix, it went into the
infinite-loop. After fix, it can print values in the list
in reverse order.

fn main() {
    let mut list = List::new();

    list.push(100);
    list.push(110);

    list.unshift(10);
    list.unshift(20);

    let mut ll = list.foot;

    loop {
        match ll.take() {
            None => break,
            Some(cur) => {
                let cb = cur.borrow();
                println!("{}", cb.data);
                match &cb.prev {
                    None => ll = None,
                    Some(prev) => {
                        ll = prev.upgrade();
                    }
                }
            }
        }
    }
}

Signed-off-by: Gioh Kim <gurugio@gurugio.com>